intern_projects/Smart-Travel-Assistant/requirements.txt
pymongo
duckduckgo-search
requests
typing_extensions
groq
python-dotenv 
yagmail
streamlit